+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning][SemVer].
 
 ## [Unreleased]
+
+## [1.2.0] - 2020-08-16
 ### Added
 - Add experimental support for modules and include a `mod_example` with the
-  source code.  The APIs aren't documented yet and may change in the future.
+  source code. The APIs aren't documented yet and may change in the future.
 - Include `mod_stats_influx`, a module for logging STUN/TURN events/statistics
   to InfluxDB, contributed by Marc Schink.
 
@@ -122,7 +124,8 @@ project adheres to [Semantic Versioning][SemVer].
 ### Added
 - Initial (pre-)release of the eturnal STUN/TURN server.
 
-[Unreleased]: https://github.com/processone/eturnal/compare/1.1.0...HEAD
+[Unreleased]: https://github.com/processone/eturnal/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/processone/eturnal/releases/tag/1.2.0
 [1.1.0]: https://github.com/processone/eturnal/releases/tag/1.1.0
 [1.0.0]: https://github.com/processone/eturnal/releases/tag/1.0.0
 [0.8.0]: https://github.com/processone/eturnal/releases/tag/0.8.0